+++ packages/VimeoPreviewImage/VimeoPreviewImageElement.js
@@ -73,13 +73,21 @@
                 this,
                 canReadAttributes([{
                     name: 'data-video-id',
-                    validate: value => !!value,
+                    validate: (value) => !!value,
                     property: 'videoID',
+                }, {
+                    name: 'data-video-width',
+                    transform: (value) => parseInt(value, 10),
+                    validate: (value) => !value || !Number.isNaN(parseInt(value, 10)),
+                    property: 'width',
                 }]),
             );
             this.readAttributes();
         }
 
+        /**
+        * Use async so that we can await the api call in our tests
+        */
         async connectedCallback() {
             await this.updateImageSource();
         }
@@ -97,7 +105,14 @@
         }
 
         getVideoInfoURL() {
-            return `${this.baseURL}${this.urlInfoPath}?url=${this.baseURL}${this.videoID}`;
+            const params = {
+                url: `${this.baseURL}${this.videoID}`,
+                ...(this.width ? { width: this.width } : {})
+            };
+            const getParams = Object.entries(params)
+                .map(([key, value]) => `${key}=${value}`)
+                .join('&');
+            return `${this.baseURL}${this.urlInfoPath}?${getParams}`;
         }
 
         /**